# Spring III/fp-be-java-hisp-w15-g06sbazzano/docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"fp-be-java-hisp-w15-g06","description":"fp-be-java-hisp-w15-g06","contact":{"name":"ext_marmarqu","email":"ext_marmarqu@mercadolibre.com"},"version":"last"},"servers":[{"url":"http://localhost:8080","description":"Scope local"}],"paths":{"/api/v1/fresh-products/orders/{idOrder}":{"get":{"tags":["product-controller"],"operationId":"findProductsByOrder","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"idOrder","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"put":{"tags":["product-controller"],"operationId":"updateOrder","parameters":[{"name":"idOrder","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}}}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["inbound-order-controller"],"operationId":"updateInboundOrder","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}},"post":{"tags":["inbound-order-controller"],"operationId":"saveInboundOrder","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}}},"/api/v1/sign-in":{"post":{"tags":["session-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/api/v1/fresh-products/orders":{"post":{"tags":["product-controller"],"operationId":"createOrder","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SuccessfulOrderDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/{idProduct}/warehouse/list":{"get":{"tags":["product-controller"],"operationId":"findAllProductsInWarehouse","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"idProduct","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductWarehouseDTO"}}}}}}},"/api/v1/fresh-products/{idProduct}/batch/list":{"get":{"tags":["stock-controller"],"operationId":"findAllBatchByProduct","parameters":[{"name":"idProduct","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockResponseDTO"}}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["product-controller"],"operationId":"findProduct_1","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"category","in":"query","required":true,"schema":{"type":"string","enum":["FS","RF","FF"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}},"/api/v1/fresh-products/batch/list/due-date/{cantDays}":{"get":{"tags":["due-date-stock-controller"],"operationId":"dueDateStockProducts","parameters":[{"name":"cantDays","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","enum":["FS","RF","FF"]}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":"date_asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockListResponseDTO"}}}}}}}},"components":{"schemas":{"PurchaseOrderDTO":{"type":"object","properties":{"date":{"type":"string","format":"date"},"products":{"type":"array","items":{"$ref":"#/components/schemas/RequestProductDTO"}},"buyer_id":{"type":"integer","format":"int64"},"order_status":{"type":"string","enum":["SHOPPING_CART"]}}},"RequestProductDTO":{"type":"object","properties":{"quantity":{"type":"integer","format":"int32"},"product_id":{"type":"integer","format":"int64"}}},"BatchStockDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product_id":{"type":"integer","format":"int64"},"current_temperature":{"type":"number","format":"double"},"minimum_temperature":{"type":"number","format":"double"},"initial_quantity":{"type":"integer","format":"int32"},"current_quantity":{"type":"integer","format":"int32"},"manufacturing_time":{"type":"string","format":"date-time"},"due_date":{"type":"string","format":"date"},"inbound_id":{"type":"integer","format":"int64"}}},"InboundOrderDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"order_date":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectionDTO"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockDTO"}}}},"InboundOrderRequestDTO":{"type":"object","properties":{"inbound_order":{"$ref":"#/components/schemas/InboundOrderDTO"}}},"SectionDTO":{"type":"object","properties":{"section_code":{"type":"integer","format":"int64"},"warehouse_code":{"type":"integer","format":"int64"}}},"InboundOrderResponseDTO":{"type":"object","properties":{"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockDTO"}}}},"UserRequestDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"name":{"type":"string"},"token":{"type":"string"}}},"SuccessfulOrderDTO":{"type":"object","properties":{"total_price":{"type":"number","format":"double"}}},"ProductWarehouseDTO":{"type":"object","properties":{"product_id":{"type":"integer","format":"int64"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/WarehouseQuatityResponseDTO"}}}},"WarehouseQuatityResponseDTO":{"type":"object","properties":{"warehouse_code":{"type":"integer","format":"int64"},"total_quantity":{"type":"integer","format":"int32"}}},"BatchStockResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product_id":{"type":"integer","format":"int64"},"current_temperature":{"type":"number","format":"double"},"minimum_temperature":{"type":"number","format":"double"},"initial_quantity":{"type":"integer","format":"int32"},"current_quantity":{"type":"integer","format":"int32"},"manufacturing_time":{"type":"string","format":"date-time"},"due_date":{"type":"string","format":"date"}}},"ProductDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"BatchStockListResponseDTO":{"type":"object","properties":{"batch_stock_dtolist":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockDTO"}}}}}}}