# Spring III/fp-be-java-hisp-w15-g06sbazzano/src/main/java/com/mercadolibre/fp_be_java_hisp_w15_g06/dtos/InboundOrderRequestDTO.java
package com.mercadolibre.fp_be_java_hisp_w15_g06.dtos;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
public class InboundOrderRequestDTO {

    private InboundOrderDTO inbound_order;
}