# Spring III/fp-be-java-hisp-w15-g06sbazzano/src/main/java/com/mercadolibre/fp_be_java_hisp_w15_g06/dtos/InboundOrderResponseDTO.java
package com.mercadolibre.fp_be_java_hisp_w15_g06.dtos;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;

@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
public class InboundOrderResponseDTO {
    private List<BatchStockDTO> batch_stock;

}